% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesRdata-package.R
\docType{data}
\name{nes}
\alias{nes}
\title{National Eutrophication Survey Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 775 rows and 38 columns.}
\usage{
nes
}
\description{
A dataset containing hydrologic and water quality data for approximately 800
lakes in the continental United States.
}
\keyword{datasets}
