% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nes_get}
\alias{nes_get}
\title{nes_get}
\usage{
nes_get(version_id, dest_folder = tempdir(), skip = NA, compile = TRUE)
}
\arguments{
\item{version_id}{character version id}

\item{dest_folder}{file.path optional will default to the location returned by \code{\link[rappdirs]{user_data_dir}}.}

\item{skip}{numeric vector of lines to skip on file read. optional.}

\item{compile}{logical perform on-the-fly compilation to rds?}
}
\description{
Retrieves external files and store in file cache.
}
\examples{
\dontrun{
nes_get(version_id = "5") # save to temp folder

nes_get(version_id = "5", dest_folder = cache_path()) # save to cache folder
}
}
