% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{nes_ingest}
\alias{nes_ingest}
\title{Ingest flat files}
\usage{
nes_ingest(version_id, folder = NA, skip = NA)
}
\arguments{
\item{version_id}{character nes version string}

\item{folder}{file.path to data folder. optional.}

\item{skip}{numeric vector of lines to skip on file read. optional.}
}
\description{
Ingest data from component flat files
}
\examples{
\dontrun{
nes_ingest("1")
}

}
