% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nes_load}
\alias{nes_load}
\title{nes_load}
\usage{
nes_load(version_id, folder = tempdir(), format = "rds", fpath = NA)
}
\arguments{
\item{version_id}{character database version string}

\item{folder}{file.path to data folder; use cache_path() to load cached (non-temporary) data}

\item{format}{character choice of rds or sqlite}

\item{fpath}{file.path optionally specify custom location of rds file}
}
\description{
Load files from local file system
}
\examples{
\dontrun{

# load from tempdir
nes_get("5")
dt  <- nes_load("5")

# load from cached
nes_get("5", dest_folder = cache_path())
dt  <- nes_load("5")
}
}
