% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{nes_compile}
\alias{nes_compile}
\title{Compile data to R data (rds) object}
\usage{
nes_compile(version_id, format = "rds", folder = NA, skip = NA)
}
\arguments{
\item{version_id}{character temp database version string}

\item{format}{character choice of "rds" or "sqlite"}

\item{folder}{file.path to data folder.}

\item{skip}{numeric vector of lines to skip on file read. optional.}
}
\description{
Compile data from component flat files
}
\examples{
\dontrun{
nes_compile("1", format = "rds")
}
}
