% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nes_load}
\alias{nes_load}
\title{temp_load}
\usage{
nes_load(version_id, format = "rds", fpath = NA)
}
\arguments{
\item{version_id}{character database version string}

\item{format}{character choice of rds or sqlite}

\item{fpath}{file.path optionally specify custom location of rds file}
}
\description{
Load files from local file system
}
\examples{
\dontrun{
dt  <- nes_load("1")
dt2 <- nes_load("2")
}
}
