% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{nes_ls}
\alias{nes_ls}
\title{nes_ls}
\usage{
nes_ls(version_id, folder = temp_path(), ...)
}
\arguments{
\item{version_id}{character version id}

\item{folder}{file.path to NES data; pass cache_path() to use OS agnostic cache location specified by the rappdirs package.}

\item{...}{extra arguments passed to list.files}
}
\description{
nes_ls
}
\examples{
nes_ls("1")
}
